MaddTheSane: - Clean up the Xcode projects a bit. - Added 64-bit mode on Mac. - Made it so that the SDK was set on the Xcode projects. - Enabled PowerPC support. (Patch #7730).
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@61171 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/plugins/DFInput/macsrc/MappingCell.m
@@ -31,13 +31,13 @@
 	return self;
 }
 
-- (void)selectWithFrame:(NSRect)aRect inView:(NSView *)controlView editor:(NSText *)textObj delegate:(id)anObject start:(int)selStart length:(int)selLength
+- (void)selectWithFrame:(NSRect)aRect inView:(NSView *)controlView editor:(NSText *)textObj delegate:(id)anObject start:(NSInteger)selStart length:(NSInteger)selLength
 {
 	[super selectWithFrame:aRect inView:controlView editor:textObj delegate:anObject start:selStart length:selLength];
 
-	int whichPad = [ControllerList currentController];
+	NSInteger whichPad = [ControllerList currentController];
 	NSTableView *tableView = (NSTableView *)[self controlView];
-	int i, changed = 0, row;
+	NSInteger i, changed = 0, row;
 	NSEvent *endEvent;
 	NSPoint where = {0.0, 0.0};
